requirements.txt
# Core Framework
# Keep Django at current stable version for dev/test
Django==5.0.1
python-dotenv==1.0.0

# Database
# psycopg2-binary works fine on Linux/Railway (Windows build issues are development-only)
psycopg2-binary==2.9.11
dj-database-url==2.1.0

# Async Task Queue - Email & Background Jobs
celery==5.3.4
redis==5.0.1

# Frontend
Pillow>=11.0.0
django-crispy-forms>=2.3
crispy-bootstrap5==2025.6

# Authentication
django-allauth==0.57.0

# REST API
djangorestframework==3.14.0

# Admin Interface
django-grappelli==4.0.1

# ngrok Tunneling
pyngrok==5.1.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development / CI tools
black==23.9.1
isort==5.12.0
flake8==6.1.0
pre-commit==3.4.0
